LPS-170489 Rename ServiceConfigurationExtender for ServiceConfigurationExtension. Start ServiceConfigurationExtension in LiferayServiceExtender.
